requirements.txt
autocrop==1.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dlib==19.21.1
dominate==2.6.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.20.1
opencv-python==4.5.1.48
passlib==1.7.4
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
